requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
httpx==0.28.1
sqlalchemy==2.0.48
psycopg==3.3.3
pydantic==2.12.5
python-dotenv==1.1.0
